2023-04-22 23:55 UTC-0500 Ron Pinkas <ronpinkas/AT/gmail/com>    * .gitignore      + *.log|ppo|pp$ tests/*.c      * make_drw.sh        * Added to C_USR "-Wno-invalid-source-encoding"
   * contrib/hbssl/err.c
   * contrib/hbssl/ssl.c
   * contrib/hbssl/sslciph.c
   * contrib/rdd_ads/adsfunc.c
   * source/common/hbwince.c
   * source/odbc/odbc.c
   * source/rtl/filesys.c
   * source/rtl/fstemp.c
     ! Fixed compiler warnings - Corrected char iniitialization from '\0' to 0

   * include/hbapi.h
   * source/compiler/comptool.c
   * source/pp/ppgen.c
   * source/vm/cmdarg.c
     ! Fixed compiler warning - return value of hb_verCvsID() int to long long int

   * include/set.ch
     + Added missing constants:
         #define HB_SET_TRACESTACK_NONE    0
         #define HB_SET_TRACESTACK_CURRENT 1
         #define HB_SET_TRACESTACK_ALL     2

   * source/common/hbmem.c
     * Protected with #ifndef HB_NO_DV_MEMCPY

   * source/compiler/cmdcheck.c
     * Fixed compiler warnings - apparent typo in bit masking: & ( 2 ^ 6 ) -> & ((1 << 6) - 1) !!!

   * source/compiler/harbour.sly
     * Corrected deprecated %name-prefix "hb_comp_yy" to %define api.prefix {hb_comp_yy}

   * source/ct/dattime2.c
     ! Fixed compiler warning - Apparent typo, corrected: && to || !!!

   * source/ct/disk.c
      ! Fixed compiler warning - protected windows specific code with #ifdef HB_OS_WIN

   * source/ct/files.c
     ! Fixed compiler warning - added: HB_SYMBOL_UNUSED( szFile );

   * source/ct/settime.c
     ! Fixed compiler warning - commented unused variable: fResult

   * source/macro/Makefile
     - Removed -p hb_comp_yy from YACC_FLAGS
   * source/macro/macro.slx
     * Added #define YYSTYPE HB_MACRO_YYSTYPE
   * source/macro/macro.y
     * Fixed compiler warnings - Corrected deprecated %pure-parser to %define api.pure
       and %name-prefix "hb_macro_yy" to %define api.prefix {hb_macro_yy}
   * source/macro/macroyy.c
      * Fixed compiler warnings - corrected deprecated %name-prefix "hb_macro_yy" to %define api.prefix {hb_macro_yy}

   * source/rdd/hbsix/sxtable.c
     ! Fixed compiler warning - Apparent typo, corrected: || to && !!!

   * source/rtl/filenet.c
     ! Fixed compiler warning - %hu to %u and cast to void * by interim (intptr_t) cast

   * source/rtl/gtcrs/gtcrs.c
   * source/rtl/gttrm/gttrm.c
     ! Fixed compiler warning - %hd to %d

   * source/rtl/gtxwc/gtxwc.c
     ! Fixed compiler warning - type cast to const char * and Commented un used variable: whiteColor

   * source/rtl/pcre/pcre_study.c
     ! Fixed compiler warning - Apparent typo corrected: && to & on line 1258 !!!

   * source/tip/cgi.prg
     ! Fixed compiler warning - commented out 2 unused STATIC function_exists

   * source/vm/classes.c
      ! Fixed compiler warning - Apparent typo corrected: && to & !!!

   * source/vm/hbservmt.c
     ! Fixed compiler warning - Unused variable: pStack

   * source/vm/itemapi.c
     ! Fixed compiler warning - Added:
         #elif defined(__APPLE__)
           #define HB_FINITE_DBL( d )  ( isfinite( d ) != 0 )

   * tests/tsttrace.prg
     - Removed no longer needed: #defines

   * utils/hbfilere/hbfilere.c
     ! Fixed compiler warning - wrong type of main, wrong format specifiers and missing type casts

   * utils/misc/hbrc.c
     ! Fixed compiler warning - wrong format specifiers

   * utils/xbscript/xbscript.prg
     ! Fixed bug in NextExp() - missing LOOP after ++/-- post op

   * winmake/makefile.gc
     - Removed -p hb_comp_yy from Bison flags

+++ include/hbapi.h
@@ -999,7 +999,7 @@ extern char *           hb_verPlatform( void ); /* retrieves a newly allocated b
 extern char *           hb_verCompiler( void ); /* retrieves a newly allocated buffer containing compiler version */
 extern char *           hb_verHarbour( void ); /* retrieves a newly allocated buffer containing harbour version */
 extern char *           hb_verBuildInfo( BOOL ); /* display harbour, compiler, and platform versions to standard console */
-extern int              hb_verCvsID( void ); /* ChangeLog CVS revision number */
+extern long long int    hb_verCvsID( void ); /* ChangeLog CVS revision number */
 extern int              hb_verSVNDateID( void ); /* ChangeLog Date */
 extern const char *     hb_verCvsChangeLogID( void ); /* ChangeLog ID string */
 extern const char *     hb_verCvsLastEntry( void ); /* ChangeLog last entry string */
