+++ source/vm/cmdarg.c
@@ -411,7 +411,7 @@ void hb_cmdargProcessVM( void )
 }
 
 /* ChangeLog CVS revision number */
-int hb_verCvsID( void )
+long long int hb_verCvsID( void )
 {
    return HB_VER_CVSID;
 }
